Moved a few functions to the utils file.
